Feat - Image adding on vela done

+++ inventory.manager/src/main/java/com/velas/vivene/inventory/manager/controller/ImagemController.java
@@ -1,22 +1,24 @@
 package com.velas.vivene.inventory.manager.controller;
 
 
-import com.velas.vivene.inventory.manager.dto.usuario.UsuarioResponseDto;
+import java.io.IOException;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+
 import com.velas.vivene.inventory.manager.entity.Imagem;
 import com.velas.vivene.inventory.manager.infraService.S3BucketService;
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.media.Schema;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import io.swagger.v3.oas.annotations.responses.ApiResponses;
+
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.*;
-
-import java.io.IOException;
 
 @Controller
 @RequestMapping("/imagem")
@@ -28,7 +30,7 @@ public class ImagemController {
     private final S3BucketService service;
 
     @GetMapping(value = "/foto/{id}", produces = "image/jpeg")
-    public ResponseEntity<byte[]> getFoto(@PathVariable Long id) throws IOException {
+    public ResponseEntity<byte[]> getFoto(@PathVariable Integer id) throws IOException {
 
         byte[] imagem = service.buscarFoto(id);
 
@@ -46,7 +48,7 @@ public class ImagemController {
     }
 
     @PatchMapping(value = "/foto/{id}", consumes = "image/*")
-    public ResponseEntity<Void> atualizarFoto(@PathVariable Long id, @RequestBody byte[] referenciaArquivoFoto) {
+    public ResponseEntity<Void> atualizarFoto(@PathVariable Integer id, @RequestBody byte[] referenciaArquivoFoto) {
         service.salvarFoto(id, referenciaArquivoFoto);
         return ResponseEntity.ok().build();
     }
